+++ .ci.sh
@@ -276,31 +276,36 @@ import modulePath from "path";
                 }
             },
             "defines": [
-
+                //
 // https://www.sqlite.org/compile.html#recommended_compile_time_options
-
+                //
+                "SQLITE_DQS=0",
+                // "SQLITE_THREADSAFE=0",
                 "SQLITE_DEFAULT_MEMSTATUS=0",
                 "SQLITE_DEFAULT_WAL_SYNCHRONOUS=1",
-                "SQLITE_DQS=0",
                 "SQLITE_LIKE_DOESNT_MATCH_BLOBS",
-                // "SQLITE_MAX_EXPR_DEPTH=0",
-                // "SQLITE_OMIT_AUTOINIT",
+                "SQLITE_MAX_EXPR_DEPTH=0",
                 "SQLITE_OMIT_DECLTYPE",
                 "SQLITE_OMIT_DEPRECATED",
                 "SQLITE_OMIT_PROGRESS_CALLBACK",
                 // "SQLITE_OMIT_SHARED_CACHE",
-                // "SQLITE_THREADSAFE=0",
                 "SQLITE_USE_ALLOCA",
-
-// custom
-
+                "SQLITE_OMIT_AUTOINIT",
+                //
+                // extra optimization
+                //
+                "HAVE_MALLOC_USABLE_SIZE",
+                "SQLITE_ENABLE_NULL_TRIM",
+                "SQLITE_ENABLE_SORTER_REFERENCES",
+                //
+                // extra feature
+                //
                 "SQLITE_ENABLE_MATH_FUNCTIONS",
                 // "SQLITE_ENABLE_STMTVTAB",
-                // "SQLITE_ENABLE_UNKNOWN_SQL_FUNCTION",
                 "SQLITE_HAVE_ZLIB",
-
-// node-sqlite3
-
+                //
+                // node-sqlite3
+                //
                 "HAVE_USLEEP=1",
                 "NAPI_DISABLE_CPP_EXCEPTIONS=1",
                 "SQLITE_ENABLE_DBSTAT_VTAB=1",
@@ -471,6 +476,7 @@ shCiBuildWasm() {(set -e
     local FILE2
     local OPTION1
     local OPTION2
+    printf "shCiBuildWasm\n" 1>&2
     OPTION1="$OPTION1 -I.tmp"
     OPTION1="$OPTION1 -Wall"
     OPTION1="$OPTION1 -flto"
@@ -487,11 +493,13 @@ shCiBuildWasm() {(set -e
         # optimization - skip rebuild of sqlite3.c if possible
         if [ "$FILE2" -nt "$FILE" ] && [ "$FILE" = sqlite3.c ]
         then
+            printf "shCiBuildWasm - skip $FILE\n" 1>&2
             continue
         fi
         OPTION2=""
-        # https://www.sqlite.org/compile.html
-        # #recommended_compile_time_options
+        #
+# https://www.sqlite.org/compile.html#recommended_compile_time_options
+        #
         OPTION2="$OPTION2 -DSQLITE_DQS=0"
         OPTION2="$OPTION2 -DSQLITE_THREADSAFE=0"
         OPTION2="$OPTION2 -DSQLITE_DEFAULT_MEMSTATUS=0"
@@ -504,11 +512,15 @@ shCiBuildWasm() {(set -e
         # OPTION2="$OPTION2 -DSQLITE_OMIT_SHARED_CACHE"
         # OPTION2="$OPTION2 -DSQLITE_USE_ALLOCA"
         OPTION2="$OPTION2 -DSQLITE_OMIT_AUTOINIT"
+        #
         # extra optimization
+        #
         OPTION2="$OPTION2 -DHAVE_MALLOC_USABLE_SIZE"
         OPTION2="$OPTION2 -DSQLITE_ENABLE_NULL_TRIM"
         OPTION2="$OPTION2 -DSQLITE_ENABLE_SORTER_REFERENCES"
+        #
         # extra feature
+        #
         OPTION2="$OPTION2 -DEMSCRIPTEN"
         OPTION2="$OPTION2 -DSQLITE_DISABLE_LFS"
         OPTION2="$OPTION2 -DSQLITE_ENABLE_FTS3"
@@ -529,31 +541,32 @@ shCiBuildWasm() {(set -e
         emcc $OPTION1 $OPTION2
     done
     OPTION2=""
-    OPTION2="$OPTION2 -s ASSERTIONS=1"
-    OPTION2="$OPTION2 -s SAFE_HEAP=1"
-    if [ "$1" != --debug ]
-    then
+    #
+    OPTION2="$OPTION2 -s EXPORTED_FUNCTIONS=_sqlite3_initialize"
+    OPTION2="$OPTION2,___dbFileImportOrExport"
+    OPTION2="$OPTION2,_dbClose"
+    OPTION2="$OPTION2,_dbExec"
+    OPTION2="$OPTION2,_dbFileImportOrExport"
+    OPTION2="$OPTION2,_dbNoop"
+    OPTION2="$OPTION2,_dbOpen"
+    OPTION2="$OPTION2,_jsbatonValueErrmsg"
+    OPTION2="$OPTION2,_jsbatonValueStringArgi"
+    OPTION2="$OPTION2,_sqlite3_errmsg"
+    OPTION2="$OPTION2,_sqlite3_free"
+    OPTION2="$OPTION2,_sqlite3_malloc"
+    #
+    OPTION2="$OPTION2 -s EXPORTED_RUNTIME_METHODS=cwrap"
+    # OPTION2="$OPTION2,AsciiToString"
+    #
+    case "$1" in
+    --debug)
+        OPTION2="$OPTION2 -s ASSERTIONS=1 -s SAFE_HEAP=1"
+        ;;
+    *)
         OPTION2="$OPTION2 --closure 1"
-    fi
+        ;;
+    esac
     emcc $OPTION1 $OPTION2 \
-        -s EXPORTED_FUNCTIONS='[
-            "___dbFileImportOrExport",
-            "_dbClose",
-            "_dbExec",
-            "_dbFileImportOrExport",
-            "_dbNoop",
-            "_dbOpen",
-            "_jsbatonValueErrmsg",
-            "_sqlite3_errmsg",
-            "_sqlite3_free",
-            "_sqlite3_initialize",
-            "_sqlite3_malloc"
-        ]' \
-        -s EXPORTED_RUNTIME_METHODS='[
-            "AsciiToString",
-            "cwrap"
-        ]' \
-        \
         --memory-init-file 0 \
         --pre-js sqlmath_wrapper_wasm.js \
         -Wall \
@@ -572,15 +585,20 @@ shCiBuildWasm() {(set -e
         .tmp/sqlmath_custom.c.wasm.o \
         #
     printf '' > sqlmath_wasm.js
-    printf '/*jslint-disable*/
+    printf "/*jslint-disable*/
+// Copyright (c) 2021 Kai Zhu
+// SPDX-License-Identifier: MIT
+// $(date -u +"%Y-%m-%dT%H:%M:%S%z")
 (function () {
-"use strict";
-' >> sqlmath_wasm.js
-    cat .tmp/sqlmath_wasm.js >> sqlmath_wasm.js
+\"use strict\";
+" >> sqlmath_wasm.js
+    cat .tmp/sqlmath_wasm.js | tr -d "\r" >> sqlmath_wasm.js
     printf '
 }());
 /*jslint-enable*/
 ' >> sqlmath_wasm.js
+    cp .tmp/sqlmath_wasm.wasm .
+    ls -l sqlmath_wasm.*
 )}
 
 shCiEmsdkExport() {