+++ .fndraw.js
@@ -3,9 +3,6 @@
 /*global
     DataTable
     _fnCallbackFire
-    _fnLog
-    _fnSortFlatten
-    _pluck
     assertOrThrow
     jQuery
     stringHtmlSafe
@@ -152,7 +149,7 @@ function _fnDraw(settings) {
 function _fnAddData2(settings, aDataIn) {
     /* Create the object for storing information about this new row */
     let iRow = settings.aoData.length;
-    let oData = jQuery.extend(true, {}, DataTable.models.oRow, {
+    let oData = Object.assign({}, DataTable.models.oRow, {
         idx: iRow,
         src: "data"
     });
@@ -166,10 +163,6 @@ function _fnAddData2(settings, aDataIn) {
     });
     /* Add to the display array */
     settings.aiDisplayMaster.push(iRow);
-    let id = settings.rowIdFn(aDataIn);
-    if (id !== undefined) {
-        settings.aIds[id] = oData;
-    }
     return iRow;
 }
 // _fnDraw - end